chore: renamed tagger pkg to tag
